linux: build with -Wextra
95% of this is removing "const" from return types, but turning this
on found one bug!  (A "for" loop that expected its iterator to go negative but
which was using an unsigned type.)

BUG=34160

Review URL: http://codereview.chromium.org/570012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@38266 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 225c8f507421a2eff2ed7a900104431d04ed7e5e

+++ chrome/browser/task_manager.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2009 The Chromium Authors. All rights reserved.
+// Copyright (c) 2010 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -260,7 +260,7 @@ std::pair<int, int> TaskManagerModel::GetGroupRangeForResource(int index)
   if (group->size() == 1) {
     return std::make_pair(index, 1);
   } else {
-    for (size_t i = index; i >= 0; --i) {
+    for (int i = index; i >= 0; --i) {
       if (resources_[i] == (*group)[0])
         return std::make_pair(i, group->size());
     }
